Refactor: Rename Julia backend to 'pardiso', add legacy support, and fix spin handling
- Renamed `dptb/postprocess/julia` to `dptb/postprocess/pardiso`.
- Updated `io.jl` to support legacy `.dat` files via `load_structure_dat`.
- Fixed `load_structure_json` to correctly account for spin degeneracy.
- Updated `pdso.py` and tests to reflect directory rename.
